arrange code: seperate folders HomepageCard & HomepageFeatures; custom.css: override some '--ifm-color' variables;
