+++ src/css/custom.css
@@ -6,14 +6,18 @@
 
 /* You can override the default Infima variables here. */
 :root {
-  --ifm-color-primary: #25c2a0;
-  --ifm-color-primary-dark: rgb(33, 175, 144);
-  --ifm-color-primary-darker: rgb(31, 165, 136);
-  --ifm-color-primary-darkest: rgb(26, 136, 112);
-  --ifm-color-primary-light: rgb(70, 203, 174);
-  --ifm-color-primary-lighter: rgb(102, 212, 189);
-  --ifm-color-primary-lightest: rgb(146, 224, 208);
+  --ifm-color-primary: #3850e9;
+  --ifm-color-primary-dark: rgb(25, 45, 160);
+  --ifm-color-primary-darker: rgb(14, 30, 122);
+  --ifm-color-primary-darkest: rgb(3, 13, 79);
+  --ifm-color-primary-light: rgb(126, 141, 238);
+  --ifm-color-primary-lighter: rgb(170, 179, 239);
+  --ifm-color-primary-lightest: rgb(243, 236, 247);
   --ifm-code-font-size: 95%;
+}
+
+.footer--dark {
+  --ifm-footer-background-color: rgb(10, 15, 60);
 }
 
 .docusaurus-highlight-code-line {
